# specs/001-physical-ai-textbook/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Physical AI Textbook API
  description: |
    Backend API for the Physical AI & Humanoid Robotics Textbook platform.
    Provides authentication, RAG-powered chat, and progress tracking.
  version: 1.0.0
  contact:
    name: Panaversity
    url: https://panaversity.com
servers:
  - url: https://api.physical-ai-textbook.example.com
    description: Production server
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/progress:
    get:
      tags: [progress]
      summary: Get user's progress summary
      operationId: getProgressSummary
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Progress summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressSummary'
  /api/progress/resume:
    get:
      tags: [progress]
      summary: Get resume position
      description: Returns the most recent in-progress content for "Continue Reading" feature
      operationId: getResumePosition
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Resume position
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumePosition'
        '404':
          description: No in-progress content found
  /api/progress/modules:
    get:
      tags: [progress]
      summary: Get module-level progress
      operationId: getModuleProgress
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Module progress list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModuleProgressResponse'
  /api/progress/{content_id}:
    get:
      tags: [progress]
      summary: Get progress for specific content
      operationId: getContentProgress
      security:
        - bearerAuth: []
      parameters:
        - name: content_id
          in: path
          required: true
          schema:
            type: string
          example: module-2/chapter-1
      responses:
        '200':
          description: Content progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressResponse'
        '404':
          description: No progress found for this content

    patch:
      tags: [progress]
      summary: Update progress for content
      operationId: updateContentProgress
      security:
        - bearerAuth: []
      parameters:
        - name: content_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgressUpdate'
      responses:
        '200':
          description: Progress updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProgressUpdate:
      type: object
      required: [content_type, progress_percent]
      properties:
        content_type:
          type: string
          enum: [module, chapter, exercise]
        progress_percent:
          type: number
          format: float
          minimum: 0
          maximum: 100
        scroll_position:
          type: number
          format: float
          minimum: 0
          maximum: 1
        reading_time_delta:
          type: integer
          minimum: 0
          description: Seconds to add to reading time
    ProgressResponse:
      type: object
      required: [id, content_id, content_type, status, progress_percent]
      properties:
        id:
          type: string
          format: uuid
        content_id:
          type: string
        content_type:
          type: string
          enum: [module, chapter, exercise]
        status:
          type: string
          enum: [not_started, in_progress, completed]
        progress_percent:
          type: number
          format: float
        scroll_position:
          type: number
          format: float
          nullable: true
        reading_time_seconds:
          type: integer
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
    ProgressSummary:
      type: object
      required: [user_id, modules_started, modules_completed, chapters_completed, exercises_completed, total_reading_time_seconds]
      properties:
        user_id:
          type: string
          format: uuid
        modules_started:
          type: integer
        modules_completed:
          type: integer
        chapters_completed:
          type: integer
        exercises_completed:
          type: integer
        total_reading_time_seconds:
          type: integer
        last_content_id:
          type: string
          nullable: true
        last_accessed_at:
          type: string
          format: date-time
          nullable: true
    ResumePosition:
      type: object
      required: [content_id, content_type, progress_percent]
      properties:
        content_id:
          type: string
        content_type:
          type: string
        progress_percent:
          type: number
          format: float
        scroll_position:
          type: number
          format: float
          nullable: true
    ModuleProgressResponse:
      type: object
      required: [module_id, total_chapters, completed_chapters, total_exercises, completed_exercises, overall_progress]
      properties:
        module_id:
          type: string
        total_chapters:
          type: integer
        completed_chapters:
          type: integer
        total_exercises:
          type: integer
        completed_exercises:
          type: integer
        overall_progress:
          type: number
          format: float
        first_accessed_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true